requirements.txt
streamlit
pytesseract
Pillow
pandas
PyYAML